+++ lib/tcp-client/version.rb
@@ -1,3 +1,3 @@
 class TCPClient
-  VERSION = '0.1.5'.freeze
+  VERSION = '0.2.0'.freeze
 end
